+++ cmd/terrad/testnet.go
@@ -22,6 +22,7 @@ import (
 	"cosmossdk.io/math"
 
 	"cosmossdk.io/simapp"
+
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/tx"
